+++ common/src/main/java/fr/loudo/narrativecraft/editors/cameraangle/CameraAngleMakerEditorMaker.java
@@ -188,12 +188,13 @@ public class CameraAngleMakerEditorMaker implements EditorMaker {
         addEntityToWorld(entity, player, level, characterStory);
 
         if (entity instanceof LivingEntity livingEntity) {
-            for (int i = 0; i < characterPlacement.getItems().size(); i++) {
-                EquipmentSlot slot = EquipmentSlot.values()[i];
-                ItemStack itemStack = characterPlacement.getItems().get(i);
-                livingEntity.setItemSlot(slot, itemStack);
-                ((LivingEntityAccessor) livingEntity).callDetectEquipmentUpdates();
+            for (Map.Entry<EquipmentSlot, ItemStack> entry :
+                    characterPlacement.getItemsBySlot().entrySet()) {
+                EquipmentSlot slot = entry.getKey();
+                if (!livingEntity.canUseSlot(slot)) continue;
+                livingEntity.setItemSlot(slot, entry.getValue());
             }
+            ((LivingEntityAccessor) livingEntity).callDetectEquipmentUpdates();
         }
 
         entity.entityTags().add(ENTITY_TAG);
@@ -269,7 +270,7 @@ public class CameraAngleMakerEditorMaker implements EditorMaker {
                 characterStory,
                 position,
                 rotation,
-                new ArrayList<>(),
+                new EnumMap<>(EquipmentSlot.class),
                 true,
                 true,
                 templateReferenceId);